+++ R/table_leafletAdd.R
@@ -1,0 +1,90 @@
+#' @title Add to a leaflet interactive map for known locations
+#'
+#' @param map Leaflet map.
+#' @param locationTbl Tibble of known locations.
+#' @param extraVars Character vector of addition \code{locationTbl} column names
+#' to be shown in leaflet popups.  
+#' @param color Named color to use for map icons.
+#'
+#' @description This function adds interactive maps that will be displayed in
+#'   RStudio's 'Viewer' tab.
+#'
+#' @return A leaflet "plot" object which, if not assigned, is rendered in
+#' Rstudio's 'Viewer' tab.
+#'
+#' @rdname table_leafletAdd
+#' @export 
+#' @importFrom MazamaCoreUtils stopIfNull
+#' @importFrom sp SpatialPointsDataFrame
+#' @importFrom leaflet leaflet setView addProviderTiles addCircleMarkers
+
+table_leafletAdd <- function(
+  map = NULL,
+  locationTbl = NULL,
+  extraVars = NULL,
+  color = "red"
+) {
+
+  radius <- 10
+  opacity <- 0.5
+
+  # ----- Validate parameters --------------------------------------------------
+
+  MazamaCoreUtils::stopIfNull(map)
+  MazamaLocationUtils::validateLocationTbl(locationTbl, locationOnly = FALSE)
+
+  if ( !is.null(extraVars) ) {
+    if ( ! extraVars %in% names(locationTbl) ) {
+      stop("Variables in 'extraVars' not found in 'locationTbl'")
+    }
+  }
+
+  # ----- Create popup text ----------------------------------------------------
+
+  # Create popupText
+  popupText <- paste0(
+    "<b>", locationTbl$locationName, "</b><br>",
+    "locationID = ", locationTbl$locationID, "<br>",
+    "longitude = ", locationTbl$longitude, ", ", "latitude = ", locationTbl$latitude, "<br>",
+    "timezone = ", locationTbl$timezone, "<br>",
+    "ISO = ", locationTbl$countryCode, ".", locationTbl$stateCode, "<br>",
+    "county = ", locationTbl$county, "<br>",
+    "address = ", locationTbl$houseNumber, ", ", locationTbl$street, ", ", locationTbl$city, ", ", 
+    locationTbl$stateCode, ", ", locationTbl$zip, "<br>"
+  )
+  
+  # Add extra vars
+  for ( i in seq_along(popupText) ) {
+
+    extraText <- vector("character", length(extraVars))
+    for ( j in seq_along(extraVars) ) {
+      var <- extraVars[j]
+      extraText[j] <- paste0("<b>", var, " = ", locationTbl[i, var], "</b><br>")
+    }
+    extraText <- paste0(extraText, collapse = "")
+
+    popupText[i] <- paste0(popupText[i], extraText)
+  }
+
+  locationTbl$popupText <- popupText
+  
+  # ----- Create leaflet map ---------------------------------------------------
+  
+  m <- 
+    map %>%
+    leaflet::addCircleMarkers(
+      lng = locationTbl$longitude,
+      lat = locationTbl$latitude,
+      radius = radius,
+      fillColor = color,
+      fillOpacity = opacity,
+      stroke = FALSE,
+      popup = locationTbl$popupText,
+      layerId = locationTbl$locationID
+    )
+
+  # ----- Return ---------------------------------------------------------------
+
+  return(m)
+
+}