+++ R/table_initialize.R
@@ -12,7 +12,7 @@
 #' \item{elevation}
 #' \item{countryCode}
 #' \item{stateCode}
-#' \item{county}
+#' \item{countyName}
 #' \item{timezone}
 #' \item{houseNumber}
 #' \item{street}
@@ -49,7 +49,7 @@ table_initialize <- function() {
     "elevation" = as.numeric(NA),
     "countryCode" = as.character(NA),
     "stateCode" = as.character(NA),
-    "county" = as.character(NA),
+    "countyName" = as.character(NA),
     "timezone" = as.character(NA),
     "houseNumber" = as.character(NA),
     "street" = as.character(NA),