+++ R/utils.R
@@ -35,48 +35,30 @@ validateMazamaSpatialUtils <- function() {
 
 
 #' @title Validate longitude and latitude vectors
-#' @description Longitude and latitude vectors validated to be parseable as numeric
-#' and within the bounds -180:180 and -90:90. If validation fails, an error is
-#' generated.
-#' @param longitude Vector of longitudes in decimal degrees E.
-#' @param latitude Vector of latitudes in decimal degrees N.
-#' @param na.rm Logical specifying whether to remove \code{NA} values before
-#' validation.
-#' @return Invisibly returns \code{TRUE} if no error message has been generated.
+#'
+#' See \code{MazamaCoreUtils::\link[MazamaCoreUtils:validateLonsLats]{\%>\%}} for details.
+#'
+#' @name validateLonsLats
 #' @rdname validateLonsLats
-#' @export 
-#' 
-validateLonsLats <- function(
-  longitude = NULL,
-  latitude = NULL,
-  na.rm = FALSE
-) {
-  
-  returnVal <- MazamaCoreUtils::validateLonsLats(longitude, latitude, na.rm)
-  return(invisible(returnVal))
-  
-}
+#' @keywords internal
+#' @export
+#' @importFrom MazamaCoreUtils validateLonsLats
+#' @usage validateLonsLats(longitude = NULL, latitude = NULL, na.rm = FALSE)
+NULL
 
 
 #' @title Validate longitude and latitude values
-#' @description Longitude and latitude are validated to be parseable as numeric
-#' and within the bounds -180:180 and -90:90. If validation fails, an error is
-#' generated.
-#' @param longitude Single longitude in decimal degrees E.
-#' @param latitude Single latitude in decimal degrees N.
-#' @return Invisibly returns \code{TRUE} if no error message has been generated.
+#'
+#' See \code{MazamaCoreUtils::\link[MazamaCoreUtils:validateLonLat]{\%>\%}} for details.
+#'
+#' @name validateLonLat
 #' @rdname validateLonLat
-#' @export 
-#' 
-validateLonLat <- function(
-  longitude = NULL,
-  latitude = NULL
-) {
-  
-  returnVal <- MazamaCoreUtils::validateLonLat(longitude, latitude)
-  return(invisible(returnVal))
-  
-}
+#' @keywords internal
+#' @export
+#' @importFrom MazamaCoreUtils validateLonLat
+#' @usage validateLonLat(longitude = NULL, latitude = NULL)
+NULL
+
 
 #' @title Validate a location table
 #' @description Ensures that the incoming table has numeric \code{longitude} and